# openapi.yaml
openapi: 3.0.0
info:
  title: Localia API
  description: Localia Core API following JSON:API specification with Better Auth
  version: '1.0'
  contact:
    name: Localia Support
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/auth/get-session:
    get:
      operationId: AuthController_getSession
      summary: Get current session information
      tags:
        - Auth
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Session data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          description: No active session
          content:
            application/json:
              schema:
                type: 'null'
                example: null
...
components:
  schemas:
...
    SessionResponse:
      type: object
      properties:
        session:
          type: object
          properties:
            id:
              type: string
            expiresAt:
              type: string
              format: date-time
            token:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            ipAddress:
              type: string
            userAgent:
              type: string
            userId:
              type: string
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
              format: email
            emailVerified:
              type: boolean
            image:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set after sign-in